Docs(ui/index.mdx): FAQ Turborepo Styling Missing Solution
This commit adds guidance to a common problem a user might experience when using `Turborepo` with `Tailwindcss`. The issue is that by default in the Tailwind Config, it will look for the `node_modules` containing `fumadocs/ui` which is needed for the styling with `DocsLayout`, which explains why the root page directing the user to go to `/docs` works with tailwind as expected. The fix is to edit `tailwind.config.js` and instead of `./node_modules`, it is `../../node_modules`